backend/requirements.txt
# FastAPI и ASGI сервер
fastapi==0.110.0
uvicorn[standard]==0.27.1

# База данных
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Валидация и конфигурация
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.1
email-validator>=2.0.0

# SSH и удаленное управление
paramiko==3.4.0

# Утилиты
python-multipart==0.0.6
qrcode[pil]==7.4.2
pillow==10.2.0

# Аутентификация и безопасность
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
cryptography==42.0.0

# Планировщик задач
apscheduler==3.10.4

# HTTP клиент
httpx==0.26.0

# Rate limiting
slowapi==0.1.9

# Разработка и тестирование
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
black==24.1.1
flake8==7.0.0
